mobile homepage, pre-merging js
