+++ openlibrary/templates/home/custom_ia_carousel.html
@@ -1,9 +1,6 @@
-$def with(title, key, query=None, subject=None, work_id=None, _type=None, sorts=None, limit=None, test=False, min_books=6)
+$def with(title, key, query=None, subject=None, work_id=None, _type=None, sorts=None, limit=None, test=False, min_books=6, cta="Borrow")
 
 $ url = compose_ia_url(query=query, subject=subject, work_id=work_id, _type=_type, sorts=sorts, limit=limit, advanced=False)
-$ books = generic_carousel(key, query=query, subject=subject, work_id=work_id, _type=_type, sorts=sorts, limit=limit)
+$ books = generic_carousel(query=query, subject=subject, work_id=work_id, _type=_type, sorts=sorts, limit=limit)
 
-$if test or (books and len(books) >= min_books):
-  <div class="clearfix">
-    $:render_template("home/carousel", title, url, books, key, min_books=min_books, test=test)
-  </div>
+$:render_template("books/custom_carousel", books, title=title, url=url, key=key, test=test, min_books=min_books, cta=cta)
