+++ openlibrary/templates/lib/nav_foot.html
@@ -15,9 +15,10 @@ $def with (page)
       </ul>
     </div>
   </div>
+
   <div id="footerHistory">
     <div id="abuse" title="$_('Report a problem.')">
-      <a href="/contact?path=$request.path"><span></span>$_("Problem?")</a>
+      <a href="/contact?path=$request.path"><span></span>$_("Problems?")</a>
     </div>
   </div>
 </div>
