+++ openlibrary/templates/site/head.html
@@ -10,6 +10,11 @@
     <meta name="distribution" content="Global" />
 
     <link rel="canonical" href="$request.canonical_url" />
+
+    <link rel="stylesheet" type="text/css" href="//cdn.jsdelivr.net/jquery.slick/1.6.0/slick.css"/>
+    <link rel="stylesheet" type="text/css" href="//kenwheeler.github.io/slick/slick/slick-theme.css"/>
+    <link rel="stylesheet" type="text/css" href="//maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css"/>
+
     <link href="$static_url('build/all.css')" rel="stylesheet" type="text/css" />
 
     $if "dev" in ctx.features:
@@ -20,14 +25,24 @@
     <meta http-equiv="X-UA-Compatible" content="chrome=1" /> <!-- Enable Chrome Frame for IE -->
     <meta name="alexaVerifyID" content="wJKlTRj1Z1OI4G-J0w9R-cWhJjw" /> <!-- Necessary for Alexa -->
 
+    <script type="text/javascript" src="//code.jquery.com/jquery-1.11.0.min.js"></script>
+    <script type="text/javascript" src="//code.jquery.com/jquery-migrate-1.2.1.min.js"></script>
+    <script type="text/javascript" src="//cdn.jsdelivr.net/jquery.slick/1.6.0/slick.min.js"></script>
+    <script src="$static_url('build/new.js')" type="text/javascript"></script>
+
+    <!-- old
     <script src="/static/js/jquery-1.3.2.min.js" type="text/javascript"></script>
     <script src="/static/js/jquery-ui-1.7.2.min.js" type="text/javascript"></script>
+    -->
 
     <!-- Must be loaded after jquery -->
     <script src="//archive.org/includes/analytics.js?v=0c15bbe" type="text/javascript"></script>
 
     <script src="$static_url('build/vendor.js')" type="text/javascript"></script>
+
+    <!--
     <script src="$static_url('build/all.js')" type="text/javascript"></script>
+    -->
 
     $if query_param("_firebug") == "true":
         <script type='text/javascript' src='http://getfirebug.com/releases/lite/1.2/firebug-lite-compressed.js'></script>
