+++ openlibrary/templates/site/neck.html
@@ -16,9 +16,11 @@ $putctx("sent_neck", True)
     $for tag in ctx.get("metatags", []):
         $:tag
 
+    <!--
     <script type="text/javascript">
         \$().ready(setupSearch);
         \$().ready(linkbuttons);
     </script>
+    -->
 </head>
