+++ openlibrary/templates/site/stats.html
@@ -1,6 +1,5 @@
-<br style="clear: both"/>
 
-<div class="hidden"  id="debug-stats" style="padding: 10px 20px 20px; background: #eee; margin-bottom: 20px;">
+<div class="hidden"  id="debug-stats">
 <h2>Debug Stats</h2>
 
 <h3>SUMMARY</h3>
